Added dns lookup and affiliate link to namecheap registration page
We can't do a DNS lookup on the client side, so it has to be done on the
server side. I created a very simple express server which will do the
lookups and return the domains that are available.

Also, the express server can be used to serve up the index page to run
this locally a bit easier.

Updated the README.md to reflect new changes and how to get started